reorg and cleanup, and support for vulns/sbom scan by image path.
